Traverse history tree when building lazy states
This patch reduces the memory consumption of eager
(i.e. fully stored) path_symex_states. When such
a state is copied lazily, we traverse the tree
that represents the history of steps in order to
reconstruct the branch decisions that led to the
state. Since this can be a fairly expensive operation,
the result is cached in the state's branches field.
This is also justified by the fact that this field
is stored in contiguous memory which makes sure that
fast forwarding benefits from locality of reference.